+++ guillotina/commands/testdata.py
@@ -144,7 +144,6 @@ class TestDataCommand(Command):
             container.__name__ = self.arguments.container
             await db.async_set(self.arguments.container, container)
             await container.install()
-            self.request._container_id = container.__name__
             # Local Roles assign owner as the creator user
             roleperm = IPrincipalRoleManager(container)
             roleperm.assign_role_to_principal('guillotina.Owner', 'root')