+++ guillotina/db/writer.py
@@ -7,8 +7,8 @@ from guillotina.component import query_adapter
 from guillotina.db.interfaces import IJSONDBSerializer
 from guillotina.db.interfaces import IWriter
 from guillotina.db.orm.interfaces import IBaseObject
-from guillotina.interfaces import IContainer
 from guillotina.interfaces import IResource
+from guillotina.utils import find_container
 from guillotina.utils import get_dotted_name
 
 
@@ -20,16 +20,14 @@ class DefaultJSONDBSerializer(DefaultCatalogDataAdapter):
     Default serializer just serializer catalog data
     '''
 
-    def get_container_id(self):
-        ob = self.content
-        while getattr(ob, '__parent__', None) and not IContainer.providedBy(ob):
-            ob = ob.__parent__
-        if IContainer.providedBy(ob):
-            return ob.__name__
+    async def get_container_id(self):
+        container = await find_container()
+        if container is not None:
+            return container.__name__
 
     async def __call__(self):
         data = await super().__call__()
-        data['container_id'] = self.get_container_id()
+        data['container_id'] = await self.get_container_id()
         return data