+++ guillotina/files/utils.py
@@ -1,13 +1,15 @@
-from .const import MAX_REQUEST_CACHE_SIZE
-from guillotina.exceptions import UnRetryableRequestError
-from guillotina.utils import get_content_path
-from guillotina.utils import to_str
-
 import asyncio
 import base64
 import mimetypes
 import os
 import uuid
+
+from guillotina import task_vars
+from guillotina.exceptions import UnRetryableRequestError
+from guillotina.utils import get_content_path
+from guillotina.utils import to_str
+
+from .const import MAX_REQUEST_CACHE_SIZE
 
 
 async def read_request_data(request, chunk_size):
@@ -93,7 +95,7 @@ def guess_content_type(content_type, filename):
 
 def generate_key(request, context):
     return '{}{}/{}::{}'.format(
-        request._container_id,
+        task_vars.container.get().id,
         get_content_path(context),
         context.__uuid__,
         uuid.uuid4().hex)