+++ guillotina/utils/__init__.py
@@ -14,9 +14,11 @@ from .content import navigate_to  # noqa
 from .content import valid_id  # noqa
 from .crypto import get_jwk_key  # noqa
 from .crypto import secure_passphrase  # noqa
-from .misc import apply_coroutine  # noqa
+from .misc import apply_coroutine  # noqa; noqa
+from .misc import find_container  # noqa
 from .misc import get_current_request  # noqa
 from .misc import get_random_string  # noqa
+from .misc import get_registry  # noqa
 from .misc import get_schema_validator  # noqa
 from .misc import get_url  # noqa
 from .misc import lazy_apply  # noqa